requirements.txt
requests
pandas
python-dotenv
boto3
sqlalchemy
pymysql
matplotlib
seaborn
pyspark
nbformat
nbclient
ipykernel
